+++ bindings/python/release.sh
@@ -16,7 +16,7 @@ set -o xtrace   # Write all commands first to stderr
 set -o errexit  # Exit the script with error if any of the commands fail
 
 # The libmongocrypt git revision release to embed in our wheels.
-REVISION=$(git rev-list -n 1 1.8.1)
+REVISION=$(git rev-list -n 1 1.8.4)
 # The libmongocrypt release branch.
 BRANCH="r1.8"
 # The python executable to use.
@@ -127,10 +127,9 @@ if [ $(command -v docker) ]; then
         test_dist dist/*.whl
     fi
 
-    # TODO: requires adding a rhel-70-arm64 variant to the "upload-all" task.
-    # Build the manylinux2014 aarch64 wheel.
-    # get_libmongocrypt rhel-70-arm64 libmongocrypt/nocrypto/lib/libmongocrypt.so
-    # build_manylinux_wheel quay.io/pypa/manylinux2014_aarch64:2023-12-05-e9f0345
+    # Build the manylinux_2_28 aarch64 wheel.
+    get_libmongocrypt rhel-82-arm64 libmongocrypt/nocrypto/lib64/libmongocrypt.so
+    build_manylinux_wheel quay.io/pypa/manylinux_2_28_aarch64:2024-01-01-0e91b08
 fi
 
 ls -ltr dist